requirements.txt
asgiref
dj-database-url
Django
django-cors-headers
django-environ
django-filter
djangorestframework
environ
Markdown
pillow
psycopg2
psycopg2-binary
sqlparse
typing_extensions
tzdata
